Fix Cassandra#stop() causes OSError (at destructing object)

+++ src/testing/cassandra.py
@@ -235,7 +235,7 @@ class Cassandra(object):
             return
 
         from shutil import rmtree
-        if self._use_tmpdir:
+        if self._use_tmpdir and os.path.exists(self.base_dir):
             rmtree(self.base_dir)
 
     def read_log(self):
